Rename qsys input plugin
