+++ plugins/inputs/all/all.go
@@ -125,7 +125,7 @@ import (
 	_ "github.com/influxdata/telegraf/plugins/inputs/procstat"
 	_ "github.com/influxdata/telegraf/plugins/inputs/prometheus"
 	_ "github.com/influxdata/telegraf/plugins/inputs/puppetagent"
-	_ "github.com/influxdata/telegraf/plugins/inputs/qsys_qrc"
+	_ "github.com/influxdata/telegraf/plugins/inputs/qsc_qsys"
 	_ "github.com/influxdata/telegraf/plugins/inputs/rabbitmq"
 	_ "github.com/influxdata/telegraf/plugins/inputs/raindrops"
 	_ "github.com/influxdata/telegraf/plugins/inputs/redis"
